requirements.txt
praw
requests
python-dotenv